@Earlopain 2025-01-25
:thinking: how do I get the oldest rubocop version tests to pass when it doesn't know about `ruby34`?