@Earlopain 2025-01-30
I just added it manually for now, can be removed when the minimum supported rubocop gets bumped.